// frontend/src/types/exampleRunCreate.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * StrideTrack API
 * OpenAPI spec version: 0.1.0
 */

export interface ExampleRunCreate {
  /**
   * @minLength 1
   * @maxLength 255
   */
  athlete_name: string;
  /** @exclusiveMinimum 0 */
  distance_meters: number;
  /** @exclusiveMinimum 0 */
  duration_seconds: number;
  avg_ground_contact_time_ms?: number | null;
}